feat: improve record audit ux
Improves the record audit UI and UX to require explicit user interaction before queries will run. Audit queries tend to be slow, so it's much easier to navigate around if we allow the user to change several query parameters before actually running the search, as opposed to trying to re-fetch every time a parameter is changed. The only exception to this is pagination, which will always run a new query using the current parameters.